src/requirements.txt
Django==1.8
django-crispy-forms==1.4.0
django-registration-redux==1.1
djangorestframework==3.4.0
